pyproject.toml
[project]
name = "charity-nepal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "cookiecutter>=2.6.0",
    "crispy-tailwind>=1.0.3",
    "django>=5.2.5",
    "django-browser-reload>=1.18.0",
    "django-cors-headers>=4.7.0",
    "django-crispy-forms>=2.4",
    "django-tailwind>=4.2.0",
    "django-unfold>=0.65.0",
    "django-widget-tweaks>=1.5.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "nltk>=3.9.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "python-decouple>=3.8",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
    "statsmodels>=0.14.5",
]
